+++ src/Grid/grid.jl
@@ -700,51 +700,16 @@ struct PathOrientationInfo
     regular::Bool # Indicator whether the orientation is regular or inverted.
 end
 
-"""
-    SurfaceOrientationInfo
-
-Orientation information for 2D entities. Such an entity can be
-possibly flipped (i.e. the defining vertex order is reverse to the
-spanning vertex order) and the vertices can be rotated against each other.
-Take for example the faces
-```
-1---2 2---3
-| A | | B |
-4---3 1---4
-```
-which are rotated against each other by 90° (shift index is 1) or the faces
-```
-1---2 2---1
-| A | | B |
-4---3 3---4
-```
-which are flipped against each other. Any combination of these can happen.
-The combination to map this local face to the defining face is encoded with
-this data structure via ``rotate \\circ flip`` where the rotation is indiced by
-the shift index.
-    !!!NOTE TODO implement me.
-"""
-struct SurfaceOrientationInfo
-    #flipped::Bool
-    #shift_index::Int
+function PathOrientationInfo(edgenodes::NTuple{2, Int})
+    return PathOrientationInfo(get_edge_direction(edgenodes) > 0)
 end
 
-
 @doc raw"""
-    InterfaceOrientationInfo
+    SurfaceOrientationInfo
 
-Orientation information for 1D and 2D entities.
-The orientation is defined by the indices of the grid nodes
-associated to the vertices. To give an example, the oriented path
-```
-1 ---> 2
-```
-is called *regular*, indicated by `flipped=false`, while the oriented path
-```
-2 ---> 1
-```
-is called *inverted*, indicated by `flipped=true`.
+Orientation information for 2D entities.
 
+The orientation is defined by the indices of the grid nodes associated to the vertices.
 2D entities can be flipped (i.e. the defining vertex order is reverse to the
 spanning vertex order) and the vertices can be rotated against each other.
 
@@ -759,7 +724,7 @@ Take for example the faces
 |    \      |    \
 2-----3     3-----1
 ```
-which are rotated against each other by 240° after tranfroming to an
+which are rotated against each other by 240° after transforming to an
 equilateral triangle (shift index is 2). Or the faces
 ```
 3           2
@@ -771,20 +736,16 @@ equilateral triangle (shift index is 2). Or the faces
 ```
 which are flipped against each other.
 """
-struct OrientationInfo
+struct SurfaceOrientationInfo
     flipped::Bool
     shift_index::Int
 end
 
-function OrientationInfo(edgenodes::NTuple{2, Int})
-    return OrientationInfo(get_edge_direction(edgenodes) < 0, 0)
-end
-
-function OrientationInfo(facenodes::NTuple{N, Int}) where {N}
+function SurfaceOrientationInfo(facenodes::NTuple{N, Int}) where {N}
     min_idx = argmin(facenodes)
     shift_index = min_idx - 1
     flipped = get_face_direction(facenodes) < 0
-    return OrientationInfo(flipped, shift_index)
+    return SurfaceOrientationInfo(flipped, shift_index)
 end
 
 function get_edge_direction(edgenodes::NTuple{2, Int})